Adding support for trim operations. Closes #116
Trim operations are now supported adding /trim to the URL.
Usage: /trim:[reference pixel]:[tolerance]
Both parameters are optional.
reference pixel: top-left or bottom-right (Default: top-left)
tolerance: Integer value used as tolerance (Default: 0)

The euclidian distance between the colors of the reference pixel
and the surrounding pixels is used. If the distance is within the
tolerance they'll get trimmed. For a RGB image the tolerance would
be within the range 0-442