test_requirements.txt: Float Pillow up to (but not including) 3.0
The current Pillow version is 2.6.0 (released 2014-10-01 [1]), and
2.4.0 is stable on Gentoo.  Instead of being very conservative here,
assume Pillow uses semantic versioning [2] and will bump to 3.x if
they break backwards compatibility.  Folks who are skeptical of their
local Pillow can use the test suite to verify it, and we can blacklist
any versions that we know are bad.

It's a bit stickier if Pillow is broken for some version on only a
subset of systems, such as Pillow < 2.3.0 with FreeType >= 2.5.1
[3,4].  In that case, I'd probably just up the minimum requirement to
2.3.0, but I'm also ok leaving it up to the local packagers to sort
these things out (and shame on the FreeType folks for changing their
header installation in a patch release!).

[1]: https://pypi.python.org/pypi/Pillow/2.6.0
[2]: http://semver.org/
[3]: https://github.com/thumbor/thumbor/issues/248
[4]: https://github.com/python-pillow/Pillow/issues/425