Fill filter tested and covered.
This was a really tough one, since it required fit-in to be enabled as
well as particular dimensions to be used.

Also images must have the same size (that includes number of channels),
so I had to convert both fixture and filtered images to RGBA. Just
leaving this here if someone wonders why the changes were made to the
base test.