Fix source image download race condition
If multiple requests need the same source image around the same time
and it hasn't been downloaded them, many of those requests might
load that content at the same time, which can be very wasteful in terms
of bandwidth and time.

By locking on the URL, we solve that race condition and ensure that
the content is only downloaded once, with all dependent requests
resuming their processing as soon as the content is available to them.